Extract PaymentContext from CheckoutContext
