+++ src/Sylius/Behat/Resources/config/suites/ui/shipping/applying_shipping_fee.yml
@@ -28,5 +28,6 @@ default:
 
                 - sylius.behat.context.ui.shop.cart
                 - sylius.behat.context.ui.checkout
+                - sylius.behat.context.ui.checkout.payment
             filters:
                 tags: "@applying_shipping_fee && @ui"
